factor random optimizer out
- allows easily implementing any new random path finders as parallel, anytime algorithms
- RandomOptimizer is now the base class and RandomGreedy the specific version with random, greedy trials
- BranchOptimizer is now BranchBound which better reflects the algorithm
